Django-ecommerce/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
django-filter==23.3
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7

